fix lagi

+++ src/posts/posts.controller.ts
@@ -13,6 +13,9 @@ export class PostsController {
 
   @Get(':id')
   async findOne(@Param('id') id: string): Promise<PostModel> {
+    if (!id || id.length !== 24) {
+      throw new BadRequestException('Invalid Post ID');
+    }
     return this.postsService.findOne(id);
   }
 
@@ -29,11 +32,17 @@ export class PostsController {
     if (!post.userId) {
       throw new BadRequestException('User ID is required');
     }
+    if (!id || id.length !== 24) {
+      throw new BadRequestException('Invalid Post ID');
+    }
     return this.postsService.update(id, post);
   }
 
   @Delete(':id')
   async remove(@Param('id') id: string): Promise<void> {
+    if (!id || id.length !== 24) {
+      throw new BadRequestException('Invalid Post ID');
+    }
     await this.postsService.remove(id);
   }
 }
